requirements.txt
streamlit
pandas
numpy
scikit-fuzzy
scipy
pyrebase4
